2019/6/13 By@fukami,完成了更新分数

+++ src/main/java/com/ynu/finalwork/controller/ElectCourseController.java
@@ -4,7 +4,11 @@ import com.alibaba.fastjson.JSONObject;
 import com.ynu.finalwork.entity.ElectCourse;
 import com.ynu.finalwork.service.ElectCourseService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.*;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
 
 /**
  * Created on 2019/6/12 0012
@@ -17,8 +21,12 @@ public class ElectCourseController {
     @Autowired
     ElectCourseService electCourseService;
 
+    @Value("data.DEADLINE")
+    private String DEADLINE;
+
     JSONObject jsonObject;
 
+    // 增加一条选课记录
     @PostMapping("/addRecord")
     public Object addRecord(ElectCourse ec){
         jsonObject = new JSONObject();
@@ -34,9 +42,31 @@ public class ElectCourseController {
         return jsonObject;
     }
 
+    // 删除一条选课记录
     @DeleteMapping("/deleteRecord")
-    public void deleteRecord(@RequestParam("sid") Integer sid, @RequestParam("cid") Integer cid){
-        electCourseService.deleteRecord(sid,cid);
+    public void deleteRecord(@RequestParam("sid") Integer sid, @RequestParam("cid") Integer cid) {
+        //设置日期格式
+        SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+        System.out.println(df.format(new Date()));
+        String date = df.format(new Date());
+        try {
+            Date dateNow = df.parse(date);
+            Date deadline = df.parse(DEADLINE);
+            if (dateNow.getTime() > deadline.getTime()) {
+                System.out.println("没有超过截止时间");
+                electCourseService.deleteRecord(sid, cid);
+            } else if (dateNow.getTime() <= deadline.getTime()) {
+                System.out.println("超过了截止时间，无法推选课程");
+            }
+        } catch (Exception exception) {
+            exception.printStackTrace();
+        }
+    }
+
+    // 为课程中的同学进行评分
+    @PutMapping("/updateRecord")
+    public void updateRecord(ElectCourse ec){
+        electCourseService.updateRecord(ec);
     }
 
 }
